Fixed incorrect treepath paths
Dependency checking was not working due to the lookup path being wrong.
Through we check the return value of find_library, the value for some
reason is set as a cache variable - I presume one should only use the
variable once, we re-use in the loop. Thus the values was never updated
- our condition only fails on the first result, if that was succesful,
every other check is wrong.

We explicitly unset the variable in both the true and false case - this
ensures that we can do `cmake ..` without having to delete the
CMakeCache.txt file.