Add file with macro to check for sac2c features
The idea behind introducing this is to provide macros to check for
sac2c features. This is important as some users might prefer to use
older/newer versions of the compiler, which can lead to
incompatibilities. Lacking any version checking mechanisms, this is the
alternative solution to this problem.

This file provides a single macro which checks for support of the
`#pragma header` syntax. The result is exposed via a global variable,
which is intended to be exposed via the generated `config.h` file.